Move manufacturers data layer to feature module
