+++ apps/mobile/src/main/java/com/dpanger/vehicles/HomeActivity.kt
@@ -16,7 +16,7 @@ class HomeActivity : ComponentActivity() {
         setContent {
             VehiclesTheme {
                 VehiclesApp(
-                    modifier = Modifier.fillMaxSize()
+                    modifier = Modifier.fillMaxSize(),
                 )
             }
         }
