--- data/src/main/java/com/dpanger/vehicles/data/DataModule.kt
@@ -1,14 +1,0 @@
-package com.dpanger.vehicles.data
-
-import com.dpanger.vehicles.data.manufacturers.manufacturerRepository
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-internal class DataModule {
-    @Provides
-    fun provideManufacturerRepository() = manufacturerRepository()
-}
