+++ features/makes/build.gradle.kts
@@ -15,7 +15,6 @@ android {
 }
 
 dependencies {
-    implementation(project(":data"))
     implementation(project(":ui:components"))
     implementation(project(":ui:themes"))
     implementation(project(":vehicles"))
