+++ features/manufacturers/src/main/java/com/dpanger/vehicles/features/manufacturers/di/Module.kt
@@ -1,0 +1,15 @@
+package com.dpanger.vehicles.features.manufacturers.di
+
+import com.dpanger.vehicles.features.manufacturers.data.ManufacturerRepository
+import com.dpanger.vehicles.features.manufacturers.data.ManufacturerRepositoryImpl
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+internal class Module {
+    @Provides
+    fun provideManufacturerRepository(): ManufacturerRepository = ManufacturerRepositoryImpl()
+}
