+++ features/manufacturers/src/main/java/com/dpanger/vehicles/features/manufacturers/viewmodel/ManufacturersUiState.kt
@@ -4,6 +4,8 @@ import kotlinx.collections.immutable.ImmutableList
 
 sealed interface ManufacturersUiState {
     data object Loading : ManufacturersUiState
+
     data class Success(val manufacturers: ImmutableList<UiManufacturer>) : ManufacturersUiState
+
     data object Error : ManufacturersUiState
 }
