+++ features/manufacturers/src/main/java/com/dpanger/vehicles/features/manufacturers/viewmodel/UiManufacturer.kt
@@ -2,5 +2,5 @@ package com.dpanger.vehicles.features.manufacturers.viewmodel
 
 data class UiManufacturer(
     val id: String,
-    val name: String
+    val name: String,
 )
