+++ settings.gradle.kts
@@ -16,7 +16,6 @@ dependencyResolutionManagement {
 
 rootProject.name = "Vehicles-Android"
 include(":apps:mobile")
-include(":data")
 include(":features:makes")
 include(":features:manufacturers")
 include(":ui:components")
